Moved socket polling into own files
The socket polling interface will have use elsewhere so it has been
pulled out of the socket files.
